+++ test/cypress/support/index.ts
@@ -7,6 +7,9 @@
  */
 
 import '@cypress/code-coverage/support';
+import installLogsCollector from 'cypress-terminal-report/src/installLogsCollector';
+
+installLogsCollector();
 
 /**
  * File with the helpful commands