+++ types/api/blocks.d.ts
@@ -1,4 +1,4 @@
-import {OutputData} from '../data-formats/output-data';
+import {OutputBlockData, OutputData} from '../data-formats/output-data';
 import {BlockToolData, ToolConfig} from '../tools';
 import {BlockAPI} from './block';
 
@@ -103,7 +103,6 @@ export interface Blocks {
    * @param {boolean?} needToFocus - flag to focus inserted Block
    * @param {boolean?} replace - should the existed Block on that index be replaced or not
    * @param {string} id — An optional id for the new block. If omitted then the new id will be generated
-
    */
   insert(
     type?: string,
@@ -114,6 +113,14 @@ export interface Blocks {
     replace?: boolean,
     id?: string,
   ): BlockAPI;
+
+  /**
+   * Inserts several Blocks to specified index
+   */
+  insertMany(
+    blocks: OutputBlockData[],
+    index?: number,
+  ): BlockAPI[];
 
 
   /**